ansible: fix a etcd variable dependency on docker role
docker role now needs etcd's port info as part of it's cluster url which causes
docker role to fail when it is run without the etcd role (like for cluster-control host group).
For now we setting it as a variable when running docker role for cluster-control host group to avoid
this failure.

Signed-off-by: Madhav Puri <madhav.puri@gmail.com>