separate out long running event processing into background jobs
cluster-manager processes one event at a time. Running a playbook
as part of event processing can take sometime and block subsequent requests.
It is better to fail the request instead.

This change adds a new construct called job which constitutes the smallest
possible logic to trigger a long running job i.e. it is ideal for job to only
have logic to say run a simple palybook. The bulk of (non-time-consuming)
validations shall still happen as part of event processing.

Signed-off-by: Madhav Puri <madhav.puri@gmail.com>